# backend/src/main/resources/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/vodostaji/{id}":{"get":{"tags":["vodostaji-controller"],"operationId":"getVodostajById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVodostaj"}}}}}},"put":{"tags":["vodostaji-controller"],"operationId":"updateVodostaj","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vodostaj"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVodostaj"}}}}}},"delete":{"tags":["vodostaji-controller"],"operationId":"deleteVodostaj","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/vodostaji":{"get":{"tags":["vodostaji-controller"],"operationId":"getVodostaji","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListVodostaj"}}}}}},"post":{"tags":["vodostaji-controller"],"operationId":"createVodostaj","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vodostaj"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVodostaj"}}}}}}},"/api/vodostaji/temperature/min/{min}/max/{max}":{"get":{"tags":["vodostaji-controller"],"operationId":"getVodostajiByTemperature","parameters":[{"name":"min","in":"path","required":true,"schema":{"type":"number","format":"float"}},{"name":"max","in":"path","required":true,"schema":{"type":"number","format":"float"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListVodostaj"}}}}}}},"/api/vodostaji/station/{station}":{"get":{"tags":["vodostaji-controller"],"operationId":"getVodostajiByStation","parameters":[{"name":"station","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListVodostaj"}}}}}}},"/api/vodostaji/humidity/min/{min}/max/{max}":{"get":{"tags":["vodostaji-controller"],"operationId":"getVodostajiByHumidity","parameters":[{"name":"min","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"max","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListVodostaj"}}}}}}},"/api/download/vodostaji.json":{"get":{"tags":["file-download-controller"],"operationId":"downloadJSON","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}}}}},"/api/download/vodostaji.csv":{"get":{"tags":["file-download-controller"],"operationId":"downloadCSV","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}}}}},"/api/download/schema.json":{"get":{"tags":["file-download-controller"],"operationId":"downloadSchema","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}}}}}},"components":{"schemas":{"Vjetar":{"type":"object","properties":{"smjer":{"type":"string"},"brzina":{"type":"number","format":"float"}}},"Vodostaj":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"datum":{"type":"string"},"vrijeme":{"type":"string"},"stanica":{"type":"string"},"vodostaj":{"type":"number","format":"float"},"temperatura_zraka":{"type":"number","format":"float"},"vlaznost":{"type":"integer","format":"int32"},"tlak_zraka":{"type":"integer","format":"int32"},"oborine":{"type":"number","format":"float"},"stanje_vremena":{"type":"string"},"vjetar":{"$ref":"#/components/schemas/Vjetar"}}},"ApiResponseVodostaj":{"type":"object","properties":{"status":{"type":"string"},"message":{"type":"string"},"response":{"$ref":"#/components/schemas/Vodostaj"}}},"ApiResponseListVodostaj":{"type":"object","properties":{"status":{"type":"string"},"message":{"type":"string"},"response":{"type":"array","items":{"$ref":"#/components/schemas/Vodostaj"}}}},"ApiResponseVoid":{"type":"object","properties":{"status":{"type":"string"},"message":{"type":"string"},"response":{"type":"object"}}}}}}